valkyrie cleanup 2 boogaloo & movied asteroids up

+++ NPCs/Valkyrie/Valkyrie.cs
@@ -139,7 +139,7 @@ namespace SpiritMod.NPCs.Valkyrie
 				for (int k = 0; k < npc.oldPos.Length; k++)
 				{
 					Vector2 drawPos = npc.oldPos[k] - Main.screenPosition + drawOrigin + new Vector2(0f, npc.gfxOffY);
-					Color color = npc.GetAlpha(lightColor) * (float)(((float)(npc.oldPos.Length - k) / (float)npc.oldPos.Length) / 2);
+					Color color = npc.GetAlpha(lightColor) * ((npc.oldPos.Length - k) / (float)npc.oldPos.Length / 2);
 					spriteBatch.Draw(Main.npcTexture[npc.type], drawPos, new Microsoft.Xna.Framework.Rectangle?(npc.frame), color, npc.rotation, drawOrigin, npc.scale, effects, 0f);
 				}
 			}
